+++ tsup.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/mobx.ts"],
   format: ["esm", "cjs"],
   dts: true,
   clean: true,
   sourcemap: true,
   minify: false,
   target: "node18",
-  external: ["ioredis"],
+  external: ["ioredis", "mobx"],
 });